+++ lib/vagrant-cloudstack/action/message_not_created.rb
@@ -1,16 +1,13 @@
 module VagrantPlugins
-  module AWS
+  module Cloudstack
     module Action
-      class WarnNetworks
+      class MessageNotCreated
         def initialize(app, env)
           @app = app
         end
 
         def call(env)
-          if env[:machine].config.vm.networks.length > 0
-            env[:ui].warn(I18n.t("vagrant_aws.warn_networks"))
-          end
-
+          env[:ui].info(I18n.t("vagrant_cloudstack.not_created"))
           @app.call(env)
         end
       end