+++ lib/vagrant-cloudstack/action/terminate_instance.rb
@@ -1,20 +1,20 @@
 require "log4r"
 
 module VagrantPlugins
-  module AWS
+  module Cloudstack
     module Action
       # This terminates the running instance.
       class TerminateInstance
         def initialize(app, env)
           @app    = app
-          @logger = Log4r::Logger.new("vagrant_aws::action::terminate_instance")
+          @logger = Log4r::Logger.new("vagrant_cloudstack::action::terminate_instance")
         end
 
         def call(env)
-          server = env[:aws_compute].servers.get(env[:machine].id)
+          server = env[:cloudstack_compute].servers.get(env[:machine].id)
 
           # Destroy the server and remove the tracking ID
-          env[:ui].info(I18n.t("vagrant_aws.terminating"))
+          env[:ui].info(I18n.t("vagrant_cloudstack.terminating"))
           server.destroy
           env[:machine].id = nil